% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalcules-package.R
\docType{package}
\name{animalcules-package}
\alias{animalcules}
\alias{animalcules-package}
\title{animalcules: Interactive microbiome analysis toolkit}
\description{
animalcules is an R package for utilizing up-to-date data analytics,
    visualization methods, and machine learning models to provide users
    an easy-to-use interactive microbiome analysis framework. It can be
    used as a standalone software package or users can explore their data
    with the accompanying interactive R Shiny application. Traditional
    microbiome analysis such as alpha/beta diversity and differential
    abundance analysis are enhanced, while new methods like biomarker
    identification are introduced by animalcules. Powerful interactive
    and dynamic figures generated by animalcules enable users to
    understand their data better and discover new insights.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/compbiomed/animalcules}
  \item Report bugs at \url{https://github.com/compbiomed/animalcules/issues}
}

}
\author{
\strong{Maintainer}: Yue Zhao \email{yuezh@bu.edu} (\href{https://orcid.org/0000-0001-5257-5103}{ORCID})

Authors:
\itemize{
  \item Anthony Federico \email{anfed@bu.edu} (\href{https://orcid.org/0000-0002-9200-1689}{ORCID})
  \item W. Evan Johnson \email{wej@bu.edu} (\href{https://orcid.org/0000-0002-6247-6595}{ORCID})
}

}
\keyword{internal}
