% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{counts_to_logcpm}
\alias{counts_to_logcpm}
\title{Covert a counts table to a relative abundances table}
\usage{
counts_to_logcpm(counts_table)
}
\arguments{
\item{counts_table}{A organism x sample data frame of counts}
}
\value{
A organism x sample data frame of logcpm counts
}
\description{
Covert a counts table to a relative abundances table
}
\examples{
logcpm <- counts_to_logcpm(as.data.frame(matrix(seq_len(12),4)))

}
