% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{counts_to_relabu}
\alias{counts_to_relabu}
\title{Covert a counts table to a relative abundances table}
\usage{
counts_to_relabu(counts_table)
}
\arguments{
\item{counts_table}{A organism x sample data frame of counts}
}
\value{
A organism x sample data frame of relative abundances
}
\description{
Covert a counts table to a relative abundances table
}
\examples{
counts_to_relabu(matrix(seq_len(12),4))

}
