% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_char_to_factor}
\alias{df_char_to_factor}
\title{Factorize all categorical columns}
\usage{
df_char_to_factor(df)
}
\arguments{
\item{df}{A sample x condition data frame}
}
\value{
A sample x condition data frame
}
\description{
Factorize all categorical columns
}
\examples{
df_char_to_factor(matrix(seq_len(12)))


}
