% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversities_help.R
\name{diversities_help}
\alias{diversities_help}
\title{Get alpha diversity}
\usage{
diversities_help(counts_table, index = "all", zeroes = TRUE)
}
\arguments{
\item{counts_table}{A dataframe with organism x sample}

\item{index}{one of inverse_simpson,gini_simpson,shannon,fisher,coverage,unit}

\item{zeroes}{A boolean for whether to ignore zero values}
}
\value{
A list of alpha diversity
}
\description{
Get alpha diversity
}
\examples{
diversities_help(matrix(seq_len(12), nrow = 3),index='shannon')

}
