% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_taxon_mat.R
\name{find_taxon_mat}
\alias{find_taxon_mat}
\title{Find the Taxonomy Information Matrix}
\usage{
find_taxon_mat(names, taxonLevels)
}
\arguments{
\item{names}{Row names of the taxonomy matrix}

\item{taxonLevels}{Taxon Levels of all tids}
}
\value{
taxmat Taxonomy Information Matrix
}
\description{
Find the Taxonomy Information Matrix
}
\examples{
ids <- c("ti|54005", "ti|73001", "ti|573", "ti|228277", "ti|53458")
tids <- c("54005", "73001", "573", "228277", "53458")
taxonLevels <- find_taxonomy(tids)
tax_table <- find_taxon_mat(ids, taxonLevels)

}
