% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_simpson.R
\name{gini_simpson}
\alias{gini_simpson}
\title{Get alpha diversity using gini}
\usage{
gini_simpson(x)
}
\arguments{
\item{x}{A list of counts}
}
\value{
A single value
}
\description{
Get alpha diversity using gini
}
\examples{
gini_simpson(seq_len(10))

}
