% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_tid.R
\name{grep_tid}
\alias{grep_tid}
\title{Greps the tid from the given identifier string}
\usage{
grep_tid(id)
}
\arguments{
\item{id}{Given identifier string}
}
\value{
tid string
}
\description{
Greps the tid from the given identifier string
}
\examples{
grep_tid("ti|700015|org|Coriobacterium_glomerans_PW2")
}
