% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_simpson.R
\name{inverse_simpson}
\alias{inverse_simpson}
\title{Get alpha diversity using inverse simpson}
\usage{
inverse_simpson(x)
}
\arguments{
\item{x}{A list of counts}
}
\value{
A single value
}
\description{
Get alpha diversity using inverse simpson
}
\examples{
inverse_simpson(seq_len(10))

}
