% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_categorical}
\alias{is_categorical}
\title{Check if object is categorical}
\usage{
is_categorical(v)
}
\arguments{
\item{v}{A single value}
}
\value{
Boolean
}
\description{
Check if object is categorical
}
\examples{
nums <- 2
is_categorical(nums)

}
