% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_integer0}
\alias{is_integer0}
\title{check if integer(0)}
\usage{
is_integer0(x)
}
\arguments{
\item{x}{A single value}
}
\value{
Boolean
}
\description{
check if integer(0)
}
\examples{
nums <- 2
is_integer0(nums)

}
