% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent}
\alias{percent}
\title{Format decimals to percentages}
\usage{
percent(x, digits = 2, format = "f")
}
\arguments{
\item{x}{An array of decimals}

\item{digits}{number of digits}

\item{format}{f}
}
\value{
An array of formatted strings
}
\description{
Format decimals to percentages
}
\examples{
nums <- c(0.42, 0.15, 0.4, 0.563, 0.2)
percent(nums)

}
