% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{run_animalcules}
\alias{run_animalcules}
\title{Run animalcules shiny app}
\usage{
run_animalcules(dev = FALSE)
}
\arguments{
\item{dev}{Run the applicaiton in developer mode}
}
\value{
The shiny app will open
}
\description{
Run animalcules shiny app
}
\examples{
\dontrun{
run_animalcules()
}
}
