% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\title{Get alpha diversity using shannon}
\usage{
shannon(x)
}
\arguments{
\item{x}{A list of counts}
}
\value{
A single value
}
\description{
Get alpha diversity using shannon
}
\examples{
shannon(seq_len(10))

}
