% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson_index.R
\name{simpson_index}
\alias{simpson_index}
\title{Get alpha diversity using simpson}
\usage{
simpson_index(x)
}
\arguments{
\item{x}{A list of counts}
}
\value{
A single value
}
\description{
Get alpha diversity using simpson
}
\examples{
simpson_index(seq_len(10))

}
